feat: bottom sheet, vote buttons and legacy and modern select neurons
