+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingCard.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import type { ProposalInfo, Vote } from "@dfinity/nns";
   import { votableNeurons as getVotableNeurons } from "@dfinity/nns";
-  import { getContext, onDestroy, SvelteComponent } from "svelte";
+  import { getContext, onDestroy } from "svelte";
   import { registerVotes } from "../../../services/proposals.services";
   import { i18n } from "../../../stores/i18n";
   import { definedNeuronsStore } from "../../../stores/neurons.store";
   import { votingNeuronSelectStore } from "../../../stores/proposals.store";
   import CardInfo from "../../ui/CardInfo.svelte";
   import VotingConfirmationToolbar from "./VotingConfirmationToolbar.svelte";
-  import CastVoteCardNeuronSelect from "./VotingNeuronSelect.svelte";
+  import VotingNeuronSelect from "./VotingNeuronSelect.svelte";
   import {
     SELECTED_PROPOSAL_CONTEXT_KEY,
     type SelectedProposalContext,
@@ -19,7 +19,7 @@
     type VoteInProgress,
   } from "../../../stores/voting.store";
   import { VOTING_UI } from "../../../constants/environment.constants";
-  import ContentCell from "../../ui/ContentCell.svelte";
+  import { BottomSheet } from "@dfinity/gix-components";
 
   export let proposalInfo: ProposalInfo;
 
@@ -75,19 +75,28 @@
     votingNeuronSelectStore.reset();
   });
 
-  // TODO(L2-965): delete legacy component <CardInfo />, inline styles (.content-cell-title and .content-cell-details) and delete ContentCell
-  let cmp: typeof SvelteComponent =
-    VOTING_UI === "legacy" ? CardInfo : ContentCell;
+  // TODO(L2-965): delete legacy component <CardInfo />
 </script>
 
 {#if visible}
-  <svelte:component this={cmp}>
-    <h2 slot="start">{$i18n.proposal_detail__vote.headline}</h2>
-    <CastVoteCardNeuronSelect {proposalInfo} {voteInProgress} />
-    <VotingConfirmationToolbar
-      {proposalInfo}
-      {voteInProgress}
-      on:nnsConfirm={vote}
-    />
-  </svelte:component>
+  {#if VOTING_UI === "legacy"}
+    <CardInfo>
+      <h2 slot="start">{$i18n.proposal_detail__vote.headline}</h2>
+      <VotingNeuronSelect {proposalInfo} {voteInProgress} />
+      <VotingConfirmationToolbar
+        {proposalInfo}
+        {voteInProgress}
+        on:nnsConfirm={vote}
+      />
+    </CardInfo>
+  {:else}
+    <BottomSheet>
+      <VotingConfirmationToolbar
+        {proposalInfo}
+        {voteInProgress}
+        on:nnsConfirm={vote}
+      />
+      <VotingNeuronSelect {proposalInfo} {voteInProgress} />
+    </BottomSheet>
+  {/if}
 {/if}
